+++ common/src/main/kotlin/org/valkyrienskies/clockwork/content/logistics/gas/duct/INodeBlock.kt
@@ -20,28 +20,7 @@ import org.valkyrienskies.clockwork.kelvin.api.nodes.PipeDuctNode
 import org.valkyrienskies.mod.common.util.toJOMLD
 
 interface INodeBlock : IDuct {
-    fun _createBlockStateDefinition(builder: StateDefinition.Builder<Block, BlockState>) {
-        builder.add(
-            NORTH_CONNECTION,
-            EAST_CONNECTION,
-            SOUTH_CONNECTION,
-            WEST_CONNECTION,
-            UP_CONNECTION,
-            DOWN_CONNECTION,
-            GAS_HEAT_LEVEL
-        )
-    }
 
-    fun ductConnectionsDefault(defaultBlockState: BlockState): BlockState {
-        return defaultBlockState
-            .setValue(NORTH_CONNECTION, DuctConnectionType.NONE)
-            .setValue(EAST_CONNECTION, DuctConnectionType.NONE)
-            .setValue(SOUTH_CONNECTION, DuctConnectionType.NONE)
-            .setValue(WEST_CONNECTION, DuctConnectionType.NONE)
-            .setValue(UP_CONNECTION, DuctConnectionType.NONE)
-            .setValue(DOWN_CONNECTION, DuctConnectionType.NONE)
-            .setValue(GAS_HEAT_LEVEL, GasHeatLevel.COOL)
-    }
 
     fun _onPlace(state: BlockState, level: Level, pos: BlockPos, oldState: BlockState, isMoving: Boolean) {
         if (!level.isClientSide) {
